♻️ Refactor LIKE const
