+++ pages/store/index.vue
@@ -199,7 +199,7 @@
 
 <script setup lang="ts">
 import { formatUnits } from 'viem'
-import { LIKE_TOKEN_DECIMALS } from '~/shared/constants'
+import { LIKE_TOKEN_DECIMALS } from '~/composables/use-likecoin-contract'
 
 const { t: $t, locale } = useI18n()
 const localeRoute = useLocaleRoute()
